requirements.txt
django
djangorestframework== 3.15.2
django-cors-headers
